+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Note: Go into settings to make sure it is correct.
 * Added anti-spam for when not forced by Guild Master.
 * Added keybindings for inviting and scanning (Found in settings, Invite Settings).
 Note: This does not overwrite any existing keybindings in WoW.
-* Players in an instance will not be messaged or invited.
+* Players in an instance will not be messaged or invited (fixed).
 * Fixed issue with welcome message not using GM settings.
 * Fixed issues with analytics not tracking black list and invited players.
 * Black listed players can be removed right away as long as a sync has not occurred.